# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ChainWise API",
    "version": "1.0.0",
    "description": "ChainWise Backend API - DeFi Risk Intelligence Platform"
  },
  "servers": [
    { "url": "https://chainwise-api-production.vercel.app", "description": "Production" }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health check",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HealthResponse" },
                "example": { "status": "ok", "timestamp": "2026-07-25T10:00:00.000Z", "database": "connected", "version": "1.0.0" }
              }
            }
          },
          "503": { "description": "Database disconnected" }
        }
      }
    },
    "/api/subscribe": {
      "post": {
        "summary": "Subscribe to DeFi risk alerts",
        "operationId": "postSubscribe",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SubscribeRequest" },
              "example": { "email": "you@example.com", "source": "website" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscribed successfully",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SubscribeResponse" } } }
          },
          "400": { "description": "Invalid email or already subscribed" }
        }
      }
    },
    "/api/analyze": {
      "post": {
        "summary": "Run an 8-layer risk analysis on a protocol contract",
        "operationId": "postAnalyze",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AnalyzeRequest" },
              "example": { "protocol_address": "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48", "chain": "ethereum" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Risk analysis result",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AnalyzeResponse" } } }
          },
          "400": { "description": "Invalid protocol_address or unsupported chain" }
        }
      }
    },
    "/api/cases": {
      "get": {
        "summary": "List published risk case studies",
        "operationId": "listCases",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "chain", "in": "query", "schema": { "type": "string" }, "description": "Filter by chain, e.g. ethereum" },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["risk_score", "published_date"], "default": "risk_score" } },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of published case studies",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CaseListResponse" } } }
          }
        }
      }
    },
    "/api/cases/{id}": {
      "get": {
        "summary": "Get full details for a published case study",
        "operationId": "getCase",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Full case study details including analysis breakdown and HTML report",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CaseDetail" } } }
          },
          "404": { "description": "Case not found or not published" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "database": { "type": "string" },
          "version": { "type": "string" }
        }
      },
      "SubscribeRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "mobile": { "type": "string" },
          "source": { "type": "string" }
        }
      },
      "SubscribeResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "user_id": { "type": "integer" },
          "message": { "type": "string" },
          "alerts_enabled": { "type": "boolean" }
        }
      },
      "AnalyzeRequest": {
        "type": "object",
        "required": ["protocol_address", "chain"],
        "properties": {
          "protocol_address": { "type": "string", "description": "0x-prefixed contract address" },
          "chain": { "type": "string", "enum": ["ethereum", "polygon", "arbitrum", "optimism", "base"] }
        }
      },
      "AnalysisLayer": {
        "type": "object",
        "properties": {
          "score": { "type": "integer", "minimum": 0, "maximum": 100 },
          "findings": { "type": "array", "items": { "type": "string" } }
        }
      },
      "AnalyzeResponse": {
        "type": "object",
        "properties": {
          "case_id": { "type": "integer" },
          "protocol_address": { "type": "string" },
          "chain": { "type": "string" },
          "risk_score": { "type": "integer", "minimum": 0, "maximum": 100 },
          "classification": { "type": "string", "enum": ["Normal Protocol", "MLM Pyramid Scheme", "Rug Pull Risk", "Liquidity Trap"] },
          "confidence": { "type": "number", "minimum": 0, "maximum": 1 },
          "layers": {
            "type": "object",
            "properties": {
              "patternRecognition": { "$ref": "#/components/schemas/AnalysisLayer" },
              "bytecodeAnalysis": { "$ref": "#/components/schemas/AnalysisLayer" },
              "treasuryFlow": { "$ref": "#/components/schemas/AnalysisLayer" },
              "withdrawalDynamics": { "$ref": "#/components/schemas/AnalysisLayer" },
              "bridgeExploitation": { "$ref": "#/components/schemas/AnalysisLayer" },
              "offRampActivity": { "$ref": "#/components/schemas/AnalysisLayer" },
              "sybilClustering": { "$ref": "#/components/schemas/AnalysisLayer" },
              "transactionAttribution": { "$ref": "#/components/schemas/AnalysisLayer" }
            }
          }
        }
      },
      "CaseSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "protocol_name": { "type": "string" },
          "contract_address": { "type": "string" },
          "chain": { "type": "string" },
          "risk_score": { "type": "integer" },
          "confidence": { "type": "number" },
          "classification": { "type": "string" },
          "predicted_collapse": { "type": "string" },
          "estimated_loss": { "type": "string" },
          "published_date": { "type": "string", "format": "date-time" }
        }
      },
      "CaseListResponse": {
        "type": "object",
        "properties": {
          "cases": { "type": "array", "items": { "$ref": "#/components/schemas/CaseSummary" } },
          "pagination": {
            "type": "object",
            "properties": {
              "limit": { "type": "integer" },
              "offset": { "type": "integer" },
              "total": { "type": "integer" }
            }
          }
        }
      },
      "CaseDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/CaseSummary" },
          {
            "type": "object",
            "properties": {
              "analysis": { "type": "object", "description": "8-layer risk breakdown" },
              "case_study_html": { "type": "string" },
              "published": { "type": "boolean" },
              "downloads": {
                "type": "object",
                "properties": {
                  "json": { "type": "string" },
                  "pdf": { "type": "string", "nullable": true }
                }
              }
            }
          }
        ]
      }
    }
  }
}